Change float to decimal in the Schema.sql

+++ schema.sql
@@ -5,5 +5,5 @@ CREATE TABLE animals(id integer NOT NULL default nextval('seq_id'),
  date_of_birth date,
  escape_attempts int,
  neutered boolean,
- weight_kg float,
+ weight_kg decimal,
  primary key(id));
